Remove core.editor test
This commit reverts what was done in the previous
where I added some random text to the README.
This is also going to be me trying to use better
commit etiquette from now one! Look at me using
the subject line as a subject line and the body
as the body!